move getOpponentResponse to class method
